Conflicts:
	core/src/main/java/com/adatao/ddf/content/AMetaDataHandler.java
	core/src/main/java/com/adatao/ddf/content/IHandleMetaData.java
	spark/src/main/java/com/adatao/spark/ddf/SparkDDFManager.java
	spark/src/main/java/com/adatao/spark/ddf/content/SparkRepresentationHandler.java

+++ core/src/main/java/com/adatao/ddf/content/Schema.java
@@ -122,7 +122,7 @@ public class Schema implements Serializable {
    * 
    * @return number of columns
    */
-  public long getNumColumns() {
+  public int getNumColumns() {
     return this.mColumns.size();
   }
 
@@ -202,7 +202,8 @@ public class Schema implements Serializable {
   }
 
   public enum ColumnType {
-    STRING, INTEGER, FLOAT, TIMESTAMP, BLOB;
+
+    STRING, INTEGER, DOUBLE, TIMESTAMP, BLOB;
 
     public static ColumnType fromString(String s) {
       if (s == null || s.length() == 0) return null;
